♻️ refactor(monitoring): extract page model helpers
Move Monitoring Center date, pagination, price, quota, realtime row, and scope helpers into a model module.

Keep the page focused on orchestration state and component composition while preserving behavior.

Validated with type-check, focused monitoring tests, lint, and Vite production build.
